Code and Fix -
Changed the definition of the argument "format" to be clearer to understand its role.

Corrected and moved comments to be clearer

Implemented each functions made by Xavier into each placeholder to have the appropriate result. Except for "%" where just '_putchar' is enough.

Added the incrementing of "final_size_to_print" for each loop to be able to be returned at the end of the function.